Remove genesis templates
