--- pkg/config/genesis/genesis.v3.tmpl.json
@@ -1,286 +1,0 @@
-{
-  "genesis_time": "{{ .GenesisTimeUTC }}",
-  "chain_id": "{{ .ChainID }}",
-  "initial_height": "1",
-  "consensus_params": {
-    "block": {
-      "max_bytes": "22020096",
-      "max_gas": "50000000"
-    },
-    "evidence": {
-      "max_age_num_blocks": "100000",
-      "max_age_duration": "172800000000000",
-      "max_bytes": "1048576"
-    },
-    "validator": {
-      "pub_key_types": [
-        "ed25519"
-      ]
-    },
-    "version": {}
-  },
-  "app_hash": "",
-  "app_state": {
-    "assetft": {
-      "params": {
-        "issue_fee": {
-          "denom": "{{ .Denom }}",
-          "amount": "10000000"
-        },
-        "token_upgrade_decision_timeout": "0001-01-01T00:00:00Z",
-        "token_upgrade_grace_period": "604800s"
-      }
-    },
-    "assetnft": {
-      "params": {
-        "mint_fee": {
-          "denom": "{{ .Denom }}",
-          "amount": "0"
-        }
-      }
-    },
-    "auth": {
-      "params": {
-        "max_memo_characters": "256",
-        "tx_sig_limit": "7",
-        "tx_size_cost_per_byte": "10",
-        "sig_verify_cost_ed25519": "1000",
-        "sig_verify_cost_secp256k1": "1000"
-      },
-      "accounts": []
-    },
-    "authz": {
-      "authorization": []
-    },
-    "bank": {
-      "params": {
-        "send_enabled": [],
-        "default_send_enabled": true
-      },
-      "balances": [],
-      "supply": [],
-      "denom_metadata": [
-        {
-          "description": "{{ .Denom }} coin",
-          "denom_units": [
-            {
-              "denom": "{{ .Denom }}",
-              "exponent": 0,
-              "aliases": []
-            },
-            {
-              "denom": "devcore",
-              "exponent": 6,
-              "aliases": []
-            }
-          ],
-          "base": "{{ .Denom }}",
-          "display": "devcore",
-          "name": "{{ .Denom }}",
-          "symbol": "{{ .Denom }}"
-        }
-      ]
-    },
-    "capability": {
-      "index": "1",
-      "owners": []
-    },
-    "consensus": null,
-    "crisis": {
-      "constant_fee": {
-        "denom": "{{ .Denom }}",
-        "amount": "500000000000"
-      }
-    },
-    "distribution": {
-      "params": {
-        "community_tax": "0.050000000000000000",
-        "base_proposer_reward": "0.010000000000000000",
-        "bonus_proposer_reward": "0.040000000000000000",
-        "withdraw_addr_enabled": true
-      },
-      "fee_pool": {
-        "community_pool": []
-      },
-      "delegator_withdraw_infos": [],
-      "previous_proposer": "",
-      "outstanding_rewards": [],
-      "validator_accumulated_commissions": [],
-      "validator_historical_rewards": [],
-      "validator_current_rewards": [],
-      "delegator_starting_infos": [],
-      "validator_slash_events": []
-    },
-    "evidence": {
-      "evidence": []
-    },
-    "feegrant": {
-      "allowances": []
-    },
-    "genutil": {
-      "gen_txs": []
-    },
-    "gov": {
-      "starting_proposal_id": "1",
-      "deposits": [],
-      "votes": [],
-      "proposals": [],
-      "params": {
-        "min_deposit": [
-          {
-            "denom": "{{ .Denom }}",
-            "amount": "{{ .Gov.ProposalConfig.MinDepositAmount }}"
-          }
-        ],
-        "max_deposit_period": "336h",
-        "voting_period": "{{ .Gov.ProposalConfig.VotingPeriod }}",
-        "quorum": "0.400000000000000000",
-        "threshold": "0.500000000000000000",
-        "veto_threshold": "0.334000000000000000",
-        "min_initial_deposit_ratio": "0.000000000000000000",
-        "burn_vote_quorum": false,
-        "burn_proposal_deposit_prevote": false,
-        "burn_vote_veto": true
-      }
-    },
-    "ibc": {
-      "client_genesis": {
-        "clients": [],
-        "clients_consensus": [],
-        "clients_metadata": [],
-        "params": {
-          "allowed_clients": [
-            "06-solomachine",
-            "07-tendermint"
-          ]
-        },
-        "create_localhost": false,
-        "next_client_sequence": "0"
-      },
-      "connection_genesis": {
-        "connections": [],
-        "client_connection_paths": [],
-        "next_connection_sequence": "0",
-        "params": {
-          "max_expected_time_per_block": "30000000000"
-        }
-      },
-      "channel_genesis": {
-        "channels": [],
-        "acknowledgements": [],
-        "commitments": [],
-        "receipts": [],
-        "send_sequences": [],
-        "recv_sequences": [],
-        "ack_sequences": [],
-        "next_channel_sequence": "0"
-      }
-    },
-    "mint": {
-      "minter": {
-        "inflation": "0.100000000000000000",
-        "annual_provisions": "0.000000000000000000"
-      },
-      "params": {
-        "mint_denom": "{{ .Denom }}",
-        "inflation_rate_change": "0.130000000000000000",
-        "inflation_max": "0.200000000000000000",
-        "inflation_min": "0.000000000000000000",
-        "goal_bonded": "0.670000000000000000",
-        "blocks_per_year": "17900000"
-      }
-    },
-    "monitoringp": {
-      "port_id": "monitoringp",
-      "consumerClientID": null,
-      "connectionChannelID": null,
-      "monitoringInfo": null,
-      "params": {
-        "lastBlockHeight": "1",
-        "consumerChainID": "spn-1",
-        "consumerConsensusState": {
-          "nextValidatorsHash": "",
-          "timestamp": "",
-          "root": {
-            "hash": ""
-          }
-        },
-        "consumerUnbondingPeriod": "1814400",
-        "consumerRevisionHeight": "1"
-      }
-    },
-    "params": null,
-    "slashing": {
-      "params": {
-        "signed_blocks_window": "34000",
-        "min_signed_per_window": "0.500000000000000000",
-        "downtime_jail_duration": "60s",
-        "slash_fraction_double_sign": "0.050000000000000000",
-        "slash_fraction_downtime": "0.005000000000000000"
-      },
-      "signing_infos": [],
-      "missed_blocks": []
-    },
-    "staking": {
-      "params": {
-        "unbonding_time": "168h",
-        "max_validators": 32,
-        "max_entries": 7,
-        "historical_entries": 10000,
-        "bond_denom": "{{ .Denom }}",
-        "min_commission_rate": "0.000000000000000000"
-      },
-      "last_total_power": "0",
-      "last_validator_powers": [],
-      "validators": [],
-      "delegations": [],
-      "unbonding_delegations": [],
-      "redelegations": [],
-      "exported": false
-    },
-    "transfer": {
-      "port_id": "transfer",
-      "denom_traces": [],
-      "params": {
-        "send_enabled": true,
-        "receive_enabled": true
-      }
-    },
-    "upgrade": {},
-    "vesting": {},
-    "wasm": {
-      "params": {
-        "instantiate_default_permission": "Everybody",
-        "code_upload_access": {
-          "permission": "Everybody"
-        }
-      }
-    },
-    "feemodel": {
-      "params": {
-        "model": {
-          "initial_gas_price": "0.0625",
-          "max_gas_price_multiplier": "1000.0",
-          "max_discount": "0.5",
-          "escalation_start_fraction": "0.8",
-          "max_block_gas": 50000000,
-          "short_ema_block_length": 50,
-          "long_ema_block_length": 1000
-        }
-      },
-      "min_gas_price": {
-        "denom": "{{ .Denom }}",
-        "amount": "0.0625"
-      }
-    },
-    "cnft": {},
-    "customparams": {
-      "staking_params": {
-        "min_self_delegation": "{{ .CustomParamsConfig.Staking.MinSelfDelegation }}"
-      }
-    },
-    "delay": {},
-    "group": {},
-    "nft": {}
-  }
-}
